get rid of some visual studio warnings...

+++ Moydenskall/Enumerator.cpp
@@ -48,8 +48,8 @@ void Enumerator::create_partition(std::vector<Plane>& partition, Plane& left) {
 	left.pop_back();
 
 	// insert into each existing subset
-	for (int i = 0; i < partition.size(); ++i) {
-		if (partition[i].size() < capacity) { //only if <capacity
+	for (unsigned int i = 0; i < partition.size(); ++i) {
+		if ((int) partition[i].size() < capacity) { //only if <capacity
 			partition[i].push_back(current);
 			create_partition(partition, left);
 			partition[i].pop_back();
@@ -71,7 +71,7 @@ void Enumerator::print_result(bool svg) {
 	for (auto cen : best_sites) {
 		std::cout << "FACILITY " << cen << std::endl;
 	}
-	for (int part = 0; part < best_partition.size(); ++part) {
+	for (unsigned int part = 0; part < best_partition.size(); ++part) {
 		for (auto point : best_partition[part]) {
 			std::cout << "ASSIGN " << point.ID << " " << part+1 << std::endl;
 		}
@@ -99,7 +99,7 @@ void Enumerator::svg_output() {
 
 	xmin -= 2; ymin -= 2; xmax += 2; ymax += 2;
 	double scale = 20;
-	int pointsize = ceil(std::max((xmax-xmin)/10.,(ymax-ymin)/10.));
+	int pointsize = (int) ceil(std::max((xmax-xmin)/10.,(ymax-ymin)/10.));
 
 	svgfile << "<?xml version=\"1.0\"?><!DOCTYPE svg PUBLIC \"-//W3C//DTD SVG 1.1//EN\"	\"http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd\">" << std::endl;
 	svgfile << "<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"960\" height=\"700\" viewBox = \""
@@ -110,7 +110,7 @@ void Enumerator::svg_output() {
 
 
 
-	for (int part = 0; part < best_partition.size(); ++part) {
+	for (unsigned int part = 0; part < best_partition.size(); ++part) {
 		auto cen = best_sites[part];
 		svgfile << "<circle cx = \"" << scale * cen.X << "\" cy = \"" << scale * cen.Y << "\" r = \""
 			<< pointsize << "\" style = \"fill:green\" />" << std::endl;
